+++ EventsWebApp.Application/UseCases/Users/Commands/DeleteUserCommand/DeleteUserHandler.cs
@@ -1,6 +1,7 @@
 ﻿using EventsWebApp.Application.Interfaces.UseCases;
 using EventsWebApp.Domain.Exceptions;
 using EventsWebApp.Domain.Interfaces.Repositories;
+using EventsWebApp.Domain.Models;
 
 namespace EventsWebApp.Application.UseCases.Users.Commands
 {
@@ -14,11 +15,17 @@ namespace EventsWebApp.Application.UseCases.Users.Commands
 
         public async Task<Guid> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
+            User user = await _appUnitOfWork.UserRepository.GetById(request.Id, cancellationToken);
+            if (user == null)
+            {
+                throw new UserException("No user found");
+            }
+
             int rowsDeleted = await _appUnitOfWork.UserRepository.Delete(request.Id, cancellationToken);
 
             if (rowsDeleted == 0)
             {
-                throw new SocialEventException("User wasn't deleted");
+                throw new UserException("User wasn't deleted");
             }
 
             cancellationToken.ThrowIfCancellationRequested();
